client,cli: decouple privatelink methods from common
